cyber_ids_system/requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
psutil==5.9.6
scapy==2.5.0
plotly==5.18.0
fpdf==1.7.2
mysql-connector-python==8.2.0
tensorflow==2.15.0
